Abstract option handlers
